fix: Refactor edit event button
Make onEditEvent nullable and check if not null to show button in order to fix the bug
where this button is shown in wrong places. Do not therefore provide onEditEvent value in
SearchResultSection().
Also introduces a dropdown menu with "Edit" and "Delete" options instead of the simple
"Edit Event" button to prepare for the upcoming "Delete event" action.