fix: Refactor edit event button (#386)
* fix: Refactor edit event button

Make onEditEvent nullable and check if not null to show button in order to fix the bug
where this button is shown in wrong places. Do not therefore provide onEditEvent value in
SearchResultSection().
Also introduces a dropdown menu with "Edit" and "Delete" options instead of the simple
"Edit Event" button to prepare for the upcoming "Delete event" action.

* test: Add for changes done

Add 2 tests to check edit button has been correctly removed where
unnecessary and to check the dropdownMenu has been correctly implemented

* fix: Address requested changes in PR review

Remove the ripple effect on the "more options" (three dots) icon. This is
achieved by setting `indication` to `null` and providing a new
`MutableInteractionSource` to the `.clickable` modifier.
The `onEditEvent` parameter, which was no longer in use, has also been removed from the `SearchResults` composable for code cleanup.

* fix: Format to ktfmt

---------

Co-authored-by: Eloi <eloimarie.bernier@epfl.ch>