Fix/change owner name event sheet (#415)
* feat: Fetch and display event organizer's name

This commit replaces the placeholder organizer name with the actual name fetched from the user's profile.

A new `OrganizerState` sealed class (`Loading`, `Loaded`, `Error`) has been introduced to manage the state of fetching the organizer's name asynchronously. When an event detail is opened, a request is made to the `userProfileRepository` to get the organizer's profile.

The UI in `EventDetailSheet` has been updated to react to these states, displaying "Loading...", the organizer's name, or "Unknown" in case of an error.

* test: correct previous tests and add new ones

Correct old test by changing organizerName = "..." with
 organizerState = OrganizerState.Loaded("...").
Add new tests for MapScreenViewModel to check the new functionality
Add new tests for EventDetailSheet to check the new functionality
 render properly

* fix: Address requested changes in review

Updated the KDoc and changed to OrganizerState.Error for consistency

---------

Co-authored-by: Eloi <eloimarie.bernier@epfl.ch>